requirements/vendor.txt
Markdown==2.5.2
Pygments==2.0.2
South==1.0.2
akismet==0.2.0
django-generic-aggregation==0.3.2
django-haystack==2.3.1
django-pagination==1.0.7
django-taggit==0.20.2
django-comments-spamfighter==0.4
django-recaptcha==1.0.5
django-ratelimit-backend==0.6.2
